fix: make some fixes

+++ internal/handler/interfaces.go
@@ -4,5 +4,5 @@ type CacheRecorder interface {
 	Set(key string, value []byte) error
 	Get(string) (key string, value []byte, err error)
 	Delete(key string) error
-	Size() uint
+	Size() int64
 }
