+++ pkg/storage/redis/storage.go
@@ -23,7 +23,7 @@ func New(path string) *Cache {
 
 type Cache struct {
 	client *redis.Client
-	len    uint
+	len    int64
 }
 
 func (c *Cache) Set(key string, value []byte) error {
@@ -61,6 +61,6 @@ func (c *Cache) Delete(key string) error {
 	return nil
 }
 
-func (c *Cache) Size() uint {
+func (c *Cache) Size() int64 {
 	return c.len
 }
